Cygwin: FIFO: further improve the timing when opening a writer
Add a 'wait_open_pipe' method that waits up to 50ms for a pipe
instance to become available before it tries to connect.  Use this
when opening a writer instead of open_pipe.  In the nonblocking case
return EIO if this fails.  In the blocking case keep trying.

The timeout of 50ms is just a guess as to what is needed in practice.
It may have to be adjusted.