Cygwin: FIFO: remove nreaders lock and get_nreaders
These are not needed.  They were used only to guard inc_nreaders and
dec_nreaders, which are atomic and don't need to be guarded.

Also call inc_nreaders early in open, to keep the number of readers in
sync with the status of the read_ready event.

Similarly move the call to dec_nreaders to the beginning of close.