Cygwin: FIFO: add a shared fifo_client_handler list
This is in a new shared memory section.  We will use it for temporary
storage of the owner's fc_handler list when changing owner.  The new
owner can then duplicate the pipe handles from that list before taking
ownership.

Add methods update_my_handlers and update_shared_handlers to
facilitate this process.

Allow the shared list to grow dynamically, up to a point.  Do this by
initially reserving a block of memory (currently 100 pages) and only
committing pages as needed.