Cygwin: FIFO: remove the hit_eof method
It is flawed and unnecessary.

A reader is considered to be at EOF if the pipe is empty and there are
no writers connected.  The only way to test this is to iterate through
the fifo_client_handlers as in raw_read and select.cc:peek_fifo.  Our
hit_eof was flawed because it did not do that.  Instead it relied on
the value of nconnected.  The latter didn't take into account writers
that were connected at one time but then closed.

And hit_eof is unnecessary because it was used only in raw_read and
peek_fifo, where we can just directly check EOF when we have the
necessary information.  We will do this in the next commit.

This commit also removes the nconnected data member, which was used
only for the flawed hit_eof.