Cygwin: FIFO: support opening multiple readers
Although we can have multiple readers open because of dup/fork/exec,
the current code does not support multiple readers opening a FIFO by
explicitly calling 'open'.

The main complication in supporting this is that when we open a reader
and there's already one open, we have to check the status of the
write_ready event.  The latter should be set if and only if there is
at least one writer open.

To this end we introduce a new event, check_write_ready_evt, and two
new methods check_write_ready and write_ready_checked.