Cygwin: FIFO: Handle STATUS_PIPE_CLOSING from NtFsControlFile
Occasionally NtFsControlFile(FSCTL_PIPE_LISTEN) returns
STATUS_PIPE_CLOSING.  This seems to be in the context of something
having broken earlier.  I'm not sure if the resulting pipe handle is
useable, but let's tentatively assume that it is.

I think the worst that can happen is that we later try to read from
that handle and get an error, at which time it will get declared
invalid.